requirements.txt
matplotlib==3.3.4
numpy==1.19.5
opencv_python==4.5.4.58
Pillow==9.0.1
plantcv==3.13.0
scikit_image==0.18.1
scipy==1.6.1
skimage==0.0
